Support this.transform() in before.upsert
This will make this.transform() available in before.upsert in the case
when the selector matches documents in the collection. In the case when
a new document will be inserted we can’t provide a transformed version
of the doc unless we find a way to generate said document from the
selector.